Optimize `transform_rect_bbox`
This calculates the same, but is more efficient (in part because we can
remove the affine translation to the end). This function is somewhat hot
in `Tree::update_world_recursive`.

Godbolt: https://godbolt.org/z/Y4h7P7f7P (original is
`transform_rect_bbox_foo_1`, this one is `transform_rect_bbox_foo_2`).