box_tree: Make `Tree::update_world_recursive` use a loop (#69)
This will allow for bigger trees to be updated as we're not limited by
stack size, but some of the other recursive methods may be slightly
harder to rewrite.

If the allocation turns out to be a problem, we could introduce a
reusable scratch buffer. There are some ways to do this without needing
allocation (e.g., if we have a topological order of the tree, we can
walk through the tree in that order, for each node knowing all its
ancestors have already been updated).